[SO-159] MM-bot: make markets in multiple underlyings + fix sub-dollar spot
The bot was declining DEEP/TWAL RFQs as "pair not served" — two bugs:

1. Single-pair config. Replace `underlying_symbol` with `underlying_symbols`
   (a list against the shared settlement). The bot now subscribes to every
   underlying's Pyth feed, keeps a per-underlying vol buffer, bootstraps +
   auto-replenishes inventory per underlying, and resolves each RFQ to the
   market whose pair matches the bucket — pricing against that market's spot.

2. Sub-dollar spot rounded to zero. compute_spot returned the settlement-raw-
   per-underlying-raw ratio as a u64; for DEEP (~$0.0158) and TWAL (~$0.0327)
   that ratio is < 1 and rounded to 0, so the option priced to zero. Carry the
   spot as f64 end-to-end (price_rfq / PriceDecision / mm-quote); the strike
   side already kept f64 precision.

- staging/dev/local configs quote TBTC, TDEEP, TWAL; prod left single (WBTC)
- tests: sub-unit spot precision; spot assertions/call-sites moved to f64

Stacked on SO-149 (PR #136).

+++ rust-backend/tools/mm-quote/src/main.rs
@@ -153,7 +153,7 @@ fn main() -> Result<()> {
 
 fn print_human(
     decision: &PriceDecision,
-    spot_scaled: u64,
+    spot_scaled: f64,
     inputs: &RfqPricingInputs,
     now: u64,
     args: &Args,
@@ -197,7 +197,7 @@ fn print_human(
     }
 }
 
-fn print_json(decision: &PriceDecision, spot_scaled: u64, inputs: &RfqPricingInputs, now: u64) {
+fn print_json(decision: &PriceDecision, spot_scaled: f64, inputs: &RfqPricingInputs, now: u64) {
     // Hand-roll the JSON so we don't take a serde_json dep just for this.
     print!("{{");
     print!("\"now_ms\":{now},");
